Made sprite's update width and height methods take int instead of short, however Sprite still stores the values as shorts as they will never be greater than 2048.
Also fixed an issue with movement in the set x and y methods in BaseSprite.

+++ src/main/java/org/blinz/world/Sprite.java
@@ -119,13 +119,13 @@ public abstract class Sprite extends BaseSprite {
     //PROTECTED METHODS---------------------------------------------------------
 
     @Override
-    protected final void updateWidth(final short width) {
-        this.width = width;
+    protected final void updateWidth(final int width) {
+        this.width = (short)width;
     }
 
     @Override
-    protected final void updateHeight(final short height) {
-        this.height = height;
+    protected final void updateHeight(final int height) {
+        this.height = (short)height;
     }
 
     @Override
